ctf: Allow mixing unspecified stream IDs with ID 0
If an event does not specify a stream ID, we will assign it to
stream with ID 0. This means they will be put in the same stream
as the events specifically defining stream 0.

However, the trace must not define any stream IDs other than 0.

+++ org.eclipse.tracecompass.ctf.core.tests/src/org/eclipse/tracecompass/ctf/core/tests/trace/CTFStreamInputReaderTest.java
@@ -81,7 +81,7 @@ public class CTFStreamInputReaderTest {
     private static CTFStreamInputReader getStreamInputReader() throws CTFReaderException {
         assumeTrue(testTrace.exists());
         CTFTrace trace = testTrace.getTrace();
-        CTFStream s = trace.getStream((long) 0);
+        CTFStream s = trace.getStream(0L);
         Set<CTFStreamInput> streamInput = s.getStreamInputs();
         CTFStreamInputReader retVal = null;
         for (CTFStreamInput si : streamInput) {
